# Load bloom filters on demand