--- a/table/table.go
+++ b/table/table.go
@@ -88,7 +88,6 @@ type Table struct {
 
 	fd        *os.File // Own fd.
 	tableSize int      // Initialized in OpenTable, using fd.Stat().
-	wg        sync.WaitGroup
 
 	blockIndex []*pb.BlockOffset
 	ref        int32 // For file garbage collection. Atomic.
@@ -123,7 +122,6 @@ func (t *Table) IncrRef() {
 func (t *Table) DecrRef() error {
 	newRef := atomic.AddInt32(&t.ref, -1)
 	if newRef == 0 {
-		t.wg.Wait() // Wait for the bloom filter go routine to finish.
 		// We can safely delete this file, because for all the current files, we always have
 		// at least one reference pointing to them.
 
@@ -292,7 +290,6 @@ func (t *Table) initBiggestAndSmallest() error {
 
 // Close closes the open table.  (Releases resources back to the OS.)
 func (t *Table) Close() error {
-	t.wg.Wait() // Wait for the bloom filter go routine to finish.
 	if t.opt.LoadingMode == options.MemoryMap {
 		if err := y.Munmap(t.mmap); err != nil {
 			return err
@@ -373,17 +370,6 @@ func (t *Table) readIndex() error {
 
 	t.estimatedSize = index.EstimatedSize
 	t.blockIndex = index.Offsets
-
-	// Avoid the cost of unmarshalling the bloom filters if the cache is absent.
-	if t.opt.Cache != nil {
-		t.wg.Add(1)
-		go func() {
-			defer t.wg.Done()
-			bf, err := z.JSONUnmarshal(index.BloomFilter)
-			y.Check(err)
-			t.opt.Cache.Set(t.bfCacheKey(), bf, int64(len(index.BloomFilter)))
-		}()
-	}
 	return nil
 }