@jarifibrahim 2020-03-17
This PR is no longer necessary. The issue has been fixed by https://github.com/dgraph-io/badger/pull/1260